requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.13.1
fonttools==4.55.8
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.73
ultralytics-thop==2.0.14
urllib3==2.3.0
